code/requirements.txt
transformers
sentence-transformers
faiss-cpu
pypdf
accelerate
bitsandbytes
torch
streamlit
openai
anthropic
